+++ CHANGELOG.md
@@ -14,6 +14,10 @@ All notable changes to **bUnit** will be documented in this file. The project ad
 
 - When the `TestContext` was disposed, the Blazor Renderer itself didn't dispose components under test. By [@linkdotnet](https://github.com/linkdotnet).
 
+### Added
+
+- New overloads for `IRenderedFragmentBase.InvokeAsync` that allow retrieving the work item's return value. By [@jcparkyn](https://github.com/jcparkyn).
+
 ## [1.25.3] - 2023-11-14
 
 -   Upgrade all .NET 8 preview dependencies to .NET 8 stable.